If building default (Raspberry pi), it will create a sysroot. Note the version as this needs to match what you have on your target. If you have an older image, and don't want to update you will need to edit:     cmake/rpi.sysroot.cmake
Added llvm-config call into toolchain file, which is called on the first instance of the toolchain usage.

Optimized more of the build flags.

Updated to Clang 9.0 on the compiler-rt install prefix.
I need to come up with a better solution on this one.
I experienced the 8.00 -> 9.00 change, and it will happen again...